requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
python-engineio>=4.8.0
# Produção (hospedagem)
gunicorn>=21.0.0
eventlet>=0.33.0
